logout-greeter: surface errors
previously we'd just pretend no errors occurred, which is just not cool.
surface them in the UI layer instead.

this is a bit awkward because we want this to be fairly easy to use for
third parties but also needs to be easy to create from c++. Soooo there
now is an Error.qml component that aliases properties so they are easily
set from C++ and then that Error gets created in C++ and set as
contextproperty on the greeter.

on the consuming side this allows us to simply pick up the .message (a
string) and the .action (Kirigami.Action) to present the error

talking about, our lnf now switches into an error state when an error
was detected