Always show full logout screen for "Show Logout Screen" desktop menu item
Showing the full logout screen is the intention, and it's what the
user-facing text suggests will happen.

However right now it is not always the case, because the code queries the
value of `shutdownType` in the ksmserverrc config file. This is a legacy
config key that there's no UI to set anymore, so we shouldn't be honoring
it anywhere.

BUG: 495390
FIXED-IN: 6.2.3