systemtray: Check if an icon can be located when appending symbolic
An app icon directory can either contain icon theme folder structure
or be a directory containing files to be used as icons. In the latter
case (extraSeachPath in KIconLoader speak) there is no fallback done
as there's no icon theme to do fallback in. So check if an icon can
be loaded when "symbolic" is appended before doing so.
BUG:495754