ApplicationDashboard 'Show Apps As' Config Restyling
Important: This MR is connected to

Previously, there was configuration option which enabled user to chose the way ApplicationDashboard displays application's name. Other option than AppName Only
looked clutered and was harder to read Application text.

This configuration section was removed and now application dashboard shows App Names together with
description of the application as toolTip. Before if you have chosen AppName Only, description toolTip was not visible.

For some apps, such as Konqueror, Info Center description was faulty extracted(null/empty string) in Recent Apps section which is also fixed with this MR. The reason for
that issue was that description for Recent Apps(recentusagemodel.cpp) was extracted as comment, and didn't have backup option to genericName as in appEntry.cpp -> init() function.

BUG: 459758