applets/systemtray: suppress useless tooltips
A tooltip is useless when its text is 100% identical to the visible text
of whatever you're hovering the pointer over. Let's not show them. We'll
still show the tooltip if its text is different in any way, just not if
it's identical.

BUG: 494463
FIXED-IN: 6.2.4


(cherry picked from commit 83f83a07c5a91db19244a982d870898923f9fc3d)

Co-authored-by: Nate Graham <nate@kde.org>