applets/notifications: suppress inhibited notifications after "Do not disturb" is off
After leaving "Do not disturb" mode, a flood of on-screen notifications
ensues, causing the right-hand side of the screen to be full of rapidly
scrolling and disappearing notifications for the next minute.

The user doesn't need to see any notification popup that is suppressed
by DnD mode, because the notification can be viewed from the drawer anyway.

BUG: 440837
FIXED-IN: 6.3