applets/notifications: Notification positioning fixes
Better tracking of obstructingDialog size, visibility and position,
making notifications reposition correctly when going out of edit mode.

Behavior is still a bit weird sometimes, but should never be breaking
anymore. For a full fix, the obstructingDialog should be rewritten
considering that there might be more than one, but this would
lead to a considerable increase in code complexity, so doesn't belong
in this commit.

More robust tracking of who is the parent plasmoid: the hack used before
always removed the first item from plasmoidItems which is not
guaranteed to actually be the one that was removed, now plasmoids
unsubscribe themselves before actually being destroyed, which is much
more reliable.