applets/systemtray: fix busy indicator position for hidden items
These items can also show a busy indicator, but its position was not
updated to account for the layout change in Plasma 6.2. Previously
filling/centering the busy indicator made sense with the old square
delegates, but with the new rectangular ones, that no longer makes as
much sense. Adjust the positioning accordingly.

BUG: 496910
FIXED-IN: 6.2.5