Klipper: Add a way for apps to replace items with a specific UUID
The UUID must be an actual UUID that could be read by QUuid::fromString.
Otherwise, it is ignored and the clipboard item is handled like a normal clipboard item.
This is done to prevent apps from doing something dumb like setting the UUID to "" or "1".
Technically, there is nothing preventing an app from setting a UUID like
"00000000-0000-0000-0000-000000000000", but at least that takes extra effort to do.

The UUID is converted to a SHA1 hash mainly to be consistent with other clipboard item IDs.
This also somewhat obfuscates the data files managed by klipper so that an app can't
identify its own clipboard data and manage it in a way that conflicts with Klipper.
An app could probably still get the SHA1 hash for its own generated UUID
and still mess with klipper that way, but that again takes extra effort.