Widget Explorer: re-word "uninstallable"
It's a somewhat jargonistic term that guesses at what the user is
actually wanting (to find things that can be deleted), but this is a
guess. It's safer to use clearer and more factual terminology, and let
the user figure out for themselves when using this filter mode can be
used for.

BUG: 498158
FIXED-IN: 6.3