dataengines/weather: Fix applet update on network reconnection
The weather dataengine has code to update the weather data when the
network state gets back online, by issuing a forced update via the
`reset()` method.

The ion provider stores which sources will be updated in the member:
`m_sourcesToReset = sources()`, to later emit a `forceUpdate()` signal
on those and refresh the applet.

Fix this mechanism for DWD and BBC providers, by ensuring the weather
`source` is always stored internally, which was wrongly removed in
recent refactors. Also move the relevant code to a better place and
comment appropriately so the purpose is more clear.

BUG: 498009
FIXED-IN: 6.3.0