Revert "Mark applets as disabled when they are removed from system tray"
This reverts commit 6c08876d8be5d3dd3d083a4b10988bd1f2512194.

This breaks Look and Feel themes that have a custom list of applets
for the system tray. This was needed to make the widget explorer's
"remove all instances" button also remove system tray applets, but
the preferred way forward is to not make the widget explorer handle
system tray widgets.