kcm/nightlight: Move back to "display" category
The theme category is for themes, and is the last place a user would
look for when looking for night light settings.

This was moved from display to themes in
4b84bed6e1b6c904483eded7d7770ae7fd5dddc6 implementing
https://invent.kde.org/plasma/systemsettings/-/issues/15, presumably
because Night Light was called Night Color back then and thus deemed to
be a "Color And Themes" member. But as the Night Light rename has shown,
this display-tinting feature is not seen as color customization.

This mismatch can also be seen by the fact that every single page under
themes involves selecting a theme of some sort, with Night Light being
the only exception.