kcms/wallpaper: Guard access to screen variable
The selected screen member variable is based on the user selecting a screen in the UI or on the window's current screen
either from the window changing or the screen being removed based on which happened last.

It is not guaranteed to always be valid.

BUG: 489775


(cherry picked from commit 6a8788962eee2218fe0c0319bf96a2b7130a97c5)

Co-authored-by: David Edmundson <kde@davidedmundson.co.uk>