Use tooltip title as title if no title is provided for system tray elements
Some applications, such as Discord, seem to prefer to
only set the tooltip title, and no title. This results in a
system tray entry that has the correct name when hovered,
but no name in the system tray settings. To avoid this, if
we have a tooltip title but no title, we will use the
former as the latter. In the example of Discord, this is
the only information we can use to guess the correct name.

BUG:498153