Re-arrange desktop context menu for better usability
In the past, we put the configure wallpaper item on top for
discoverability. However nowadays discoverability is provided through
System Settings, negating the need for this. The displays item was only
ever there to group it with the wallpaper item.

Accordingly, move both of them to the bottom of the menu. This lets the
"Create New" item be on top as it is in Dolphin, where it's more
contextually relevant most of the time.

BUG: 498737
FIXED-IN: 6.4