Set stepSize on seek slider to make wheel work
MPRIS operates in microseconds, so the default step size of 1 makes no
discernible difference in seeking the played track.

Set it to 5 seconds so that wheel works as expected. It also has the nice
side-effect of making the built-in left/right key handling work.

Unfortunately, it might cause tick marks to appear for short tracks.

BUG: 416128