Replace ListView timezone selector with new map type from Kirigami Components
Introduces a new TimezoneSelector component that displays a map of all timezones, a combobox, and allows you to switch between them.

This will be used in System Settings to select the system timezone and in the digital clock to select timezones to show the time of.

It uses the set of timezones provided by the system, visually represented using https://github.com/evansiroky/timezone-boundary-builder/releases and an offline map offered by openstreetmap.

See https://invent.kde.org/plasma/plasma-desktop/-/merge_requests/2572