desktopnotifier: listen trash state changes from trashrc
Currently the trash state is updated when the files in Trash/files are
changed.
Except the trashrc file that influences
KFileItem::iconFromDesktopFile called from KFileItem::iconName used to
update the trash icon, is updated in TrashImpl::fileRemoved after the Trash/files are removed and desktop notifier refreshes the trash icons.

Align the state change check with the trashrc file change, to avoid
state race-condition.