Prevent network access from the timezone map
Despite the offline tiles QtLocation still queried maps-redirect.qt.io
each time the map was opened. We don't use any of the retrieved
information, on the contrary, this is what resulted in wrong attribution/
copyright statement for the proprietary Thunderforest tiles. The tiles
are now correctly attributed to OSM as well.

While at it, also make the copyright links work.