kcms/nightlight: Fix dark theme color temperature preview
### Original commit message

* kcms/nightlight: Fix dark theme color temperature preview

The saturation adjustment for dark themes caused inconsistency in the
displayed color temperature preview. Removed the condition to ensure
accurate and consistent rendering of the preview.

### Reason for the change

In the nightlight KCM, the color day/night preview colors were not accurate for the dark theme. Specifically, temperatures above 2500K had saturation boosted. This resulted in inconsistencies if the color temperatures are close to that breakoff point. 2600K appeared warmer than 2500K. That extra condition was probably added for increased visibility of the change for cool temperatures, although I'd like to ask the original authors @bharadwaj-raju and @iasensio if there were other thoughts behind this. This change simply makes the behavior consistent across the temperature range, and also consistent with the behavior in light theme. If this is an accurate depiction of reality, I do not think we need to boost colors in any way.

### Screenshots or screen recordings

| Before | After |
| ------ | ----- |
| ![Screenshot_20250128_012932](/uploads/6553c9926b5e60c79400c7aa9c6a56a0/Screenshot_20250128_012932.png) | ![Screenshot_20250128_013549](/uploads/6c6e6477598f8bcfee9fb57852c61cb6/Screenshot_20250128_013549.png) |