applets/kicker: make submenu follow layoutDirection
When Kicker opens a cascading submenu, it always tries to open the
submenu to the right. This is contrary to the regular menu behavior in
KDE applications, where submenus open in reading direction (as specified
by layoutDirection, and against the submenu indicator arrow on the menu
item.

This change makes kicker open the submenu in reading direction first if
there is enough space.