Give ActionButton text more horizontal space
Right now ActionButtons are rather narrow, with spacing that's
inconsistent (narrow for lock+login screens, wide for logout screen).

Let's unify on narrow spacing, but expand the width of the buttons a
bit to provide more room for text. This results in almost the same
spacing visually, but increases buttons' hitboxes and makes them less
likely to have to display text in multi-line form.

BUG: 499296
FIXED-IN: 6.3.0