kcms/region_language: make removing languages work
Somewhat confusingly, the "hasImplicitLang" doesn't report whether LANG
is set implicitly, but whether both LANG and LANGUAGE are the default
value. This is tested through the default value set in the kcfg file,
which is the environment variable if it is set. Most methods changing a
value are already setting hasImplicitLang to false as part of their
operation; but remove does not. If either environment variable is set,
so is the hasImplicitLang property. Saving the languages checks for that
property, and will do nothing if it is true. Consequently, if the
environment variable is set, removing an entry from the list will remove
it from the list, but will not save this as the change is not propagated
from the selected languages model to the settings model.

This change sets hasImplicitLang to false on removal of a language, like
replacing, adding, and reordering already do.