kicker: remove favorites more reliably
previously we'd trip up on removal because the value of m_itemEntries
may no longer point to an existing resource. when that is the case we
must go by the key inside the map to relibably remove the favorite.

BUG: 497958


(cherry picked from commit 777f24cadb6dd16ad68b7a147090197c70e2bda1)

Co-authored-by: Harald Sitter <sitter@kde.org>