libnotificationmanager: don't complain about <br>
The desktop notification spec/notification portal spec don't allow
newline tags, but we explicitly support it. We are rather strict about
what we accept though; <br> by itself is valid HTML and applications
may feel inclined to use it (like Kmail, for instance, does). This
causes an issue with QXmlStreamReader, which expects that the material
follows XML rules, and we (reasonably) warn on parsing errors. If we
allow newline tags, we should handle all common cases.

This change replaces single <br> tags with <br/>. To make sure we don't
break if anyone sends <br></br> for some reason, we replace that one
first. Along the way, the handling of whitespace following a newline is
made more consistent by simplifying the detection of multiple sequential
newlines; extra whitespace is now always removed instead of only
sometimes. The new handling of newlines is also covered by an autotest.