klipper: Avoid breaking Listview
A listview by default will keep the highlight in view, scrolling the
list appropriately. This is not desired in Klipper where putting the
mouse over a clipped final delegate should not move the index.

A complicated fix was added in 31a547cc by trying to tweak at what point
we change the current index.

This is extremely messy especially when coupled with delegates of
changing heights. We start changing the current item which we know will
trigger a scroll which changes the current item. ListView gets confused
and it's not surprising.

There's a considerably simpler fix for the original bug, don't use
ListView highlights if we don't want ListView behaviour.

This fixes overlapping delegates in Klipper.

BUG: 448833