kcms/users: handle non-existing/readable profile img
kcm_users uses Avatar to display the user picture; if it can't be found,
an initial letter is used instead. The image load is attempted in Avatar
though, so an error message will be printed to stdout.

This change makes loading the image file conditional on the existing
faceValid property, avoiding unnecessary log spam in case the file can't
be read. After we fix the existing faceValid property to  actually work,
that is, by making it call Qfile::exists with a path as an argument
instead of an url.


(cherry picked from commit eab0d1b45dc281e413b191a301962c04b28013a9)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>