applets/digitalclock: only show separator with date
The new separator in Digital Clock is always active if "Always beside
time" is selected, even if the date is off. As a result, an orphaned
separator will appear in that circumstance. While this has its own logic
and can be fixed by selecting any other setting except "Always beside
time", this is not very intuitive and does not look appealing.

Instead, only show the separator when there is an actual date to
separate from the time.

BUG: 499945
FIXED-IN: 6.3.1


(cherry picked from commit 3336c0290db1e427df82731f59d89f95dfc1cf01)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>