applets/kicker: add anyRunnerFinished signal
Kicker's column-based search UI highlights the first entry in the first
visible column. As these are asynchronous, other columns need to be
aware of a new column coming in, so that the highlight can be removed
if the new column has a higher priority. However, currently each column
only has direct access to its own runner, making it cumbersome to listen
to others.

This change adds a signal that is emitted when any runner finishes to
the runner model.