weather/bbcukmet: Filter out regions from search results
The location search from bbcukmet also provides results that match
names of regions (at least within the UK). This can generate two
results with the same name for both, a town and its region.

Previously we would disambiguate those results with a number, but
this is not clear to users, so better filter out the region results.

BUG: 500065
FIXED-IN: 6.3.1