klipper: also show highlight for currentItem
Klipper's new highlight handling shows the highlight if the delegate is
hovered. The ListView's currentItem is updated as hovered is changed,
but not in the other direction, so the highlight will not correctly show
on arrow key navigation, only mouse interaction.

Instead, show the highlight on being the currentItem. As
the currentItem is updated on hover, this shows a highlight on either
keyboard or mouse interaction, and there is no chance of a double
highlight.