Move logout UI QML files to desktop package
Mixing up QML - a programming language, with themes was a bad move and we have
been moving these out of lookandfeel slowly.

We still want to allow these to be customisable, but on a per platform basis
and without the same compatibility promises.

Relevant QML file has moved to plasma-desktop.