klipper: scroll ListView on keyboard navigation
Klipper now no longer uses ListView's own highlight handling, elegantly
avoiding the issues with accidental scrolling when the mouse hovers over
items that are partially outside of view. However, this also means that
arrow key and Page Up/Down navigation no longer scrolls, whether
partially or completely outside the visible area. We need to handle this
manually if ListView does not do it for us.

This change switches to manual arrow key handling and tells the ListView
to keep the focused item completely visible, scrolling the View if
necessary (or always, in the case of Page Up/Down keys). Mouse
interaction is not affected.