applets/systemtray: Don't unnecessarily append -symbolic to icon names
41d8bb43f6697b2cd5d3fd57790b9c35d69b11a8 removed a check that the icon
name ends with "-symbolic" in favor of a function call to
`iconLoader()->hasIcon(iconName + -symbolic)`. However this introduced a
subtle bug: it will return true for icons that don't exist, if such
icons have an acceptable fallback that does exist.

This is not the behavior we want; we only want it to check the icon
theme to see if a symbolic version of the icon name exists if we know
it does not already end with -symbolic.

BUG: 500102
FIXED-IN: 6.3.2


(cherry picked from commit 78ca2a2a1a52d20ccc98644156bc5b8067e45b68)

046230ef applets/systemtray: Don't unnecessarily append -symbolic to icon names

Co-authored-by: Nate Graham <nate@kde.org>