kcms/colors: use native accent color picker
kcm_colors currently uses the QtDialogs color picker, which looks
somewhat out of place in Plasma - the buttons don't look the same, it
does not have the same features as our regular color picker does, etc.
Our regular color picker is available through Qt.labs.platform from
QML, and we already use that in a couple of places, such as the sensor
color assignment in libksysguard.

So let's also do it here.