klipper: fix clipboard actions not setting clipboard for the first time
HistoryModel::insert(QString) is for testing purpose only as the
clip data needs some time to be saved, so the overload can't be used to
set the clipboard.

Let ClipCommandProcess manually manipulate the system clipboard and notify
HistoryModel a new clip is available to update the history.

BUG: 500287
FIXED-IN: 6.3.2
(cherry picked from commit 30927e913945a36715d68c891ba09aa872b1df97)