klipper: don't hide popup when action popup shows
The klipper history menu shows a button that opens the action popup for
the selected item. This works properly in the plasmoid version of the
menu, but fails in the popup widget, as the popup hides when another
window gains focus – including the popup window it triggered, and
without it the actions popup is gone as well.

Instead, give the actions popup a recognizable name, and don't close the
history popup if the newly focused window has a matching object name.

BUG: 500438