klipper: don't scroll on Arrow Down in clear button
Arrow Down from the search bar in the klipper popup moves the
currentIndex and focuses the ListView, but the clear button right next
to it has no special handler for this. The arrow key falls back to the
ScrollView, which scrolls the view by a substantial amount. This breaks
the typical arrow key interaction pattern in the popup and feels rather
confusing.

Instead, handle it the same way as arrow down on the search field. Arrow
Up is left unchanged, as the current behavior on both controls (scolling
the view up) can be useful in case the view is not already positioned at
the top due to other user interaction. Along the way, we make sure that
the view is scrolled to the currentItem when arrowing down from either
control, so that the newly selected item is in the visible area even if
the user scrolled the view manually in between.

CCBUG: 500594