klipper: don't hide popup when action popup shows
The klipper history menu shows a button that opens the action popup for
the selected item. This works properly in the plasmoid version of the
menu, but fails in the popup widget, as the popup hides when another
window gains focus – including the popup window it triggered, and
without it the actions popup is gone as well.

Instead, give the actions popup a recognizable name, and don't close the
history popup if the newly focused window has a matching object name.

BUG: 500438


(cherry picked from commit d8a2b391a62e96a5280ab737e627ce087fd6cb69)

47c0e158 klipper: don't hide popup when action popup shows
41457203 Apply 1 suggestion(s) to 1 file(s)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>