weather/envcan: Fix parsing of station coordinates
The station coordinates in the forecast xml results are in the form
of textual "64.52N", "105.23W". We were previously only using the
absolute value, which fails for most Canada points, as "W" has to be
interpreted as a negative longitude.

The station coordinates are not exposed currently in the applet, but
are used to calculate if the observation corresponds to day or night.

BUG: 425770
FIXED-IN: 6.3.3