applets/systemtray: fix tab focus after pin button
The tab focus in the expanded representation uses a rather complex
assignment where KeyNavigation.backtab in the loaded representation is
assigned to the backtab of its container. This actually works in the
backtab direction quite well, but breaks poor KeyNavigations implicit
reverse assignment, which does not get sufficiently updated when a new
representation is loaded. Depending on the precise circumstances, it
often ends up still assigned to the previously loaded representation
which obviously can't be focused anymore, so focus just ends up stuck in
tab direction.

Instead, just set tab focus explicitly to the first entry in the
representation's focus chain. This is not always the optimal tab chain
experience by itself, but it's at least something generally reasonable
that individual applets can build on.

BUG: 500705