components/calendar: Fix previous/next buttons not always working
When you pressed the next or previous button in quick succession
it wouldn't always change the view since it waited for the animation
to end before actually changing the backend value.

Fix this by checking if we are moving through the velocity value,
and if so, trigger the next/previous in the backend.

BUG: 499871