applets/systemtray: allow fully disabling app SNIs
Some people don't like app-provided SNIs, but some apps don't offer an
option internally to disable their SNI. This is annoying.

As such, offer the option to hide app-provided SNIs from the System
Tray.

When the user does this, a message will be shown warning them that it's
at their own risk since it could break the app.

Implementation-wise, the cleanest way I found was to save the list of
disabled SNIs in a new config file key, rather than adding them to the
existing extraItems key; extraItems was designed to only hold plasmoid
IDs, and re-engineering it to accept SNI IDs as well ended up being
much more complex and fragile.

FEATURE: 489239
FEATURE: 470798
FIXED-IN: 6.4.0