applets/digital-clock: use better-looking Qt.Labs version of FontDialog
Right now we get the QML version of this dialog introduced with Qt 6.
Unfortunately it looks ugly and broken. Until Qt improves the dialog
or we override it with our own in plasma-integration, let's use the
version provided by Qt.Labs, which is the one we had for Plasma 5.

CCBUG: 487387