klipper: save MIME data to memory when "Save history across desktop sessions" is disabled
This abstracts the file read/write operations in MimeDatabase so HistoryModel and other
classes can read and write MIME data in a central place no matter where
the data is stored.

This change makes Klipper only save image data to local when "Save
history across desktop sessions" is disabled, and other data are only
kept in memory.

BUG: 501030
FIXED-IN: 6.4.0