Start ksplash on x11 after setting up environment
This way ksplash has the full environment and can read the correct
user settings. Otherwise it will ignore all look and feel defaults
because .config/kdedefaults is not inserted into XDG_CONFIG_DIRS
yet. Notably this makes it so that a look and feel theme that
specifies a different splash than itself or breeze would have no
effect.