applets/notifications: remove binding loop in popup
The notification popup sets the width of the body to equal the width
of its parent ScrollView's Flickable. This does not necessarily cause
a binding loop in a Layout, but it can if the body is long enough to
require wrapping (which esablishes the loop) but not long enough to
create a scroll bar (which would break the loop).

Instead, we bind the width to the ScrollView's width minus the scroll
bar's width, which avoids the binding loop.

BUG: 499024
FIXED-IN: 6.4.0