klipper: ignore `application/x-openoffice-link`
When accessing data of this MIME type, LibreOffice will create a DdeLink
bookmark that the user did not ask for. The correct way to disable this
behavior is for the clipboard manager ignore the MIME type data for it.

BUG: 500903
FIXED-IN: 6.3.3


(cherry picked from commit 7037a696a043f0aa7dba1fd51d4879cf43b5961d)

6089ac10 klipper: ignore `application/x-openoffice-link`
3b26c4b0 Add comment

Co-authored-by: Fushan Wen <qydwhotmail@gmail.com>