applets/systemtray: remove KSortFilterProxyModel binding loop
The systemtray's KSortFilterProxyModels bind to the model exposed on the
C++ side, which are in turn QSortFilterProxyModels. This causes a
binding loop.

The binding is unnecessary in first place though, the (sorted) models
are marked as CONSTANT and cannot themselves change (their contents can,
of course, but it's still the same sorted models). So we can just set it
once on startup and avoid the bindings and any errors or effort
associated with them.

BUG: 500374


(cherry picked from commit e3f43758f6997dee33e0d4fb7c7afcda44312a35)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>