Only make thickest visible panel reserve space for each edge
All other panels will ignore exclusive spaces on edges; this allows
to have multiple panels on the same screen edge, all touching the
screen edge instead of stacking. This was the status quo previous
to the introduction of layershells; it's not perfect (as it allows
for panels to intersect without a clear way to decide which one
should be shown above) but it's an improvement to the status quo.

BUG:477939
BUG:500668