applets/devicenotifier: handle unmounting gracefully
The deviceActions property from the beckend is briefly undefined when
a device is unmounted. As its properties are accessed to set the icon
and text of the default button and these are updated thorugh property
bindings, this causes warning messages to be printed.

Instead, guard against this with optional chaining and a reasonable
default value.