applets/devicenotifier: disable unused scrollbar
The devicenotifier popup has the ScrollBar at its default AsNeed policy,
but the content is resizing so that a horizontal scrollbar is never
needed. This is not only some extra unnecessary work, but also causes a
binding loop during initialization as the sizes of all the components
are figured out and intermediate states where a scrollbar is temporarily
marked as visible cause itself to be recomputed.

Instead, just mark the horizontal ScrollBar as AlwaysOff and avoid the
extra work and binding loop.