applets/devicenotifier: handle unmounting gracefully
The deviceActions property from the beckend is briefly undefined when
a device is unmounted. As its properties are accessed to set the icon
and text of the default button and these are updated thorugh property
bindings, this causes warning messages to be printed.

Instead, guard against this with optional chaining and a reasonable
default value.


(cherry picked from commit 4ddf61df07a674ec603c9a3eb2dea5c08b607e6b)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>