RunCommand: hide faded completion text when it matches search text
In this case, we don't need to see it since it's not actually visible.

This doesn't fully fix the issue of the text becoming bolder and/or
fuzzier (depending on font and screen scale) when the faded completion
text label is visible, but it does make the issue happen less
frequently.

CCBUG: 501208