applets/mediacontroller: keep model and TabBar index in sync
The TabBar's currentIndex is bound to the mpris model's current index.
They can still go out of sync though, in at least two instances: one
is when the user presses left or right arrow keys on the tab bar; in
contrast to other TabBars in Plasma applets (like Weather or Volume),
the content does not automatically update when the tab is selected, but
only when it is clicked or space is pressed after clicking it. The
other happens in particular during startup if
players are already running. As players get added, the currentIndex may
change from the single player to the default player, and TabBar
"helpfully" notices this and updates its own index to keep the current
one active, breaking the sync. This happens after the binding is
computed, so will remain broken until the model's currentIndex is
updated independently. These two are somewhat at odds with each other:
in case the user changes the TabBar's current index, we want the model's
currentIndex to follow, but in the case of TabBar changing the index on
its own, we want to leave it at "automatic" so it shows the application
playing media right now (unless the user specified an application
manually)

So we need to separate the two cases. We handle arrow keys by explicitly
setting setting the model's currentIndex, and letting the property
binding handle the tab bar. If the tabbar's currentIndex becomes
different from the models, we force it to stay with the model instead.