applets/mediacontroller: fix backtab in standalone widget
In the standalone version, backtab gets stuck on the tabbar. Because of
the KeyNavigation.tab set on the representation, backtab on the tabbar
tries to go there, but it does not itself have a backtab target set and
can't itself take focus, so it ends up stuck. In the tray version, this
is resolved by the tray setting a backtab target there that goes to the
tray header items (pin button etc.)

Instead, set a backtab target on the TabBar explicitly, targeting the
other end of the applet's tab chain if there's no backtab target set on
the representation.