components/widgetexplorer: add "Personal Information" category
The Merkuro contacts applet uses a "Personal information" widget
category. Widget Explorer does not support this as a translatable
category, so the string cannot be translated and a warning is printed
to this effect. This seems like a category that, while somewhat narrow,
has some applications beyond this (such as to-do widgets or
specialized calendar ones that would not fit the Date/Time category.

This change adds this as a supported translatable category.