applets/kicker: really only use forceActiveWindow on X11
When (non-dash) kicker opens a submenu, it calls forceActiveWindow() on
the submenu. This only works on X11 and causes a warning to be printed
each time on Wayland. It also does not appear to be necessary there, as
the new submenu does become active (although no items there are selected
when opened through hovering, making it not immediately obvious).

Instead, only use it on X11.