kcms/region_language: set implicitHeight on language dialog
The dialog for adding or replacing a language currently has no
implicitHeight set, taking its size from the size of its contents, which
is mainly a ListView that itself has its size determined by the
available space. This causes a binding loop with the centering logic. As
a consequence, in a addition to an unnecessary warning, the size of
the dialog changes as the user starts filtering languages, which is
an awkward experience.

Instead, set a fixed size of 80% of the parent height.