Keep notifications usable in history
This patch aims to solve the below bugs by introducing a new prop to the notification
object called `hasSomeActions`, which as the name implies, is `true` if said notification
object has any usable actions.

Upon expiry, notification popups that have actions attached to them will be kept "open"
in order to preserve functionality within the notification applet (the history), without
requiring the "resident" hint to be present.

Said notifications will still be closed upon interaction or manual dismissal by the user
(either individually or through the cleanup action), as well as expired properly when the
owning application exits. Functionality regarding the "resident" hint is untouched.


BUG: 459774
BUG: 407361
FIXED-IN: 6.4.0